improvements to addSubject
updated addSubject() function so that it takes a name, year and length
arguments. These are now added to the subjectList variable.

Also concluded getSubject, which takes an subject name (string) as an
argument and then returns the information about that subject. At the
moment thats year and length of course